# openapi.json
{"openapi":"3.1.0","info":{"title":"Indian Business Data API (Brother Edition)","version":"0.1.0"},"paths":{"/api/v1/forex/usd-inr":{"get":{"summary":"Forex","operationId":"forex_api_v1_forex_usd_inr_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/mandi/snapshot":{"get":{"summary":"Mandi Snapshot","operationId":"mandi_snapshot_api_v1_mandi_snapshot_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/ocr/pdf-to-text":{"post":{"summary":"Ocr Pdf","operationId":"ocr_pdf_api_v1_ocr_pdf_to_text_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_ocr_pdf_api_v1_ocr_pdf_to_text_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/gst/verify/{gstin}":{"get":{"summary":"Verify Gst","operationId":"verify_gst_api_v1_gst_verify__gstin__get","parameters":[{"name":"gstin","in":"path","required":true,"schema":{"type":"string","title":"Gstin"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_ocr_pdf_api_v1_ocr_pdf_to_text_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_ocr_pdf_api_v1_ocr_pdf_to_text_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}